requirements.txt
aiohttp==3.9.3
fastapi==0.110.1
INCV==0.1.1
passlib==1.7.4
pony==0.7.17
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
starlette==0.37.2
uvicorn==0.29.0